add delete , update states

+++ app/src/main/java/com/ruthvikbr/quizapp_kotlin/data/State.kt
@@ -10,14 +10,14 @@ import kotlinx.android.parcel.Parcelize
 @Entity(tableName = "State")
 data class State(
 
-
     @ColumnInfo(name = "stateName")
     var stateName: String,
 
     @ColumnInfo(name = "capitalName")
     var capitalName: String,
+
     @PrimaryKey(autoGenerate = true)
-    var StateID: Int = 0
+    var StateID: Long = 0L
 
 ) : Parcelable
